Create library application
I did it again. I made a terrible git practice, forgetting to commit
anything at all during most of the project. Therefore, this commit adds
all the main functions of the project, like adding a book, changing it's
state from read or not, and organize the cards in a grid.
Add edit and delete functions on the books, and do not forget to commit.