Fix bug on deleting a book
This commit fixes a bug where if the user deleted a book that wasn't the
last book, the others after it would be excluded from the bookLibrary
Array, this would cause an error when interacting with a card. It also
corrects a small visual error on the pages section after editing a card.